applied patchset #451 fixing warnings by Jarod42
git-svn-id: svn://maxr.org/MAX/trunk@2740 c3bd5e88-ca0b-48cf-b972-da51c77c4c1b

+++ buildings.cpp
@@ -1481,8 +1481,8 @@ bool cBuilding::canLoad ( cVehicle *Vehicle, bool checkPosition )
 
 	if ( checkPosition && !isNextTo ( Vehicle->PosX, Vehicle->PosY ) ) return false;
 
-	int i;
-	for ( i = 0; i < (int)data.storeUnitsTypes.size(); i++ )
+	size_t i;
+	for ( i = 0; i < data.storeUnitsTypes.size(); i++ )
 	{
 		if ( data.storeUnitsTypes[i].compare ( Vehicle->data.isStorageType ) == 0 ) break;
 	}
